# cache-align read-write locks for in-place memtable updates