@ajkr 2018-02-16
I'll close it for now as the measured benefit was insignificant for regular number of locks.